# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 3.0.0
  description: VulnRap.com API — Vulnerability Report Validation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
  /webhooks:
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List registered webhooks (reviewer-only)
      description: |
        Returns every registered webhook with its destination URL,
        subscribed events, registration timestamp, last successful
        delivery, and current consecutive-failure count. Reviewer-only —
        the destination URLs and failure counters are operational data we
        do not expose publicly.
      responses:
        "200":
          description: Webhook list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookList"
        "401":
          description: Missing or invalid reviewer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Register a new webhook (reviewer-only)
      description: |
        Registers a destination URL that will receive a signed POST
        whenever a subscribed event fires. v1 only emits the
        `report.scored` event after a successful POST /reports.

        The response body includes a freshly-generated `secret` exactly
        once — only its SHA-256 hash is persisted on the server, so the
        caller MUST capture this value to verify incoming HMAC
        signatures. Re-register the webhook to receive a new secret if
        it is lost.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookCreateBody"
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookCreateResponse"
        "400":
          description: Invalid URL or unsupported event type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid reviewer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /webhooks/{id}:
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Delete a registered webhook (reviewer-only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookDeleteResponse"
        "400":
          description: Invalid webhook id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid reviewer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /embed/badge.svg:
    get:
      operationId: getEmbedBadgeSvg
      tags: [embed]
      summary: Embeddable score badge SVG (5 styles)
      description: |
        Returns an SVG badge showing the slop tier + score for the report
        identified by `id` (either a numeric id or the public `VR-XXXX` hex
        code). Designed to be dropped into READMEs, advisories, HackerOne /
        Bugcrowd comments, or triage tickets via `<img>` or Markdown.

        Five visual styles are supported via the `style` query param:
        `default` (Shields-style flat with gradient), `flat`, `plastic`,
        `social`, and `square`.

        Always returns 200 OK with a well-formed SVG — unknown / hidden /
        malformed ids render an "unknown" badge so a broken image never
        appears on the embedder's page. Honours `showInFeed`: hidden
        reports render as "unknown" so private scores are not leaked. Sets
        ETag + `Cache-Control: max-age=300` so embedders that re-render
        frequently get cheap 304s.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: Numeric report id, or public `VR-XXXX` hex code.
        - name: style
          in: query
          required: false
          schema:
            type: string
            enum: [default, flat, plastic, social, square]
            default: default
      responses:
        "200":
          description: SVG badge
          content:
            image/svg+xml:
              schema:
                type: string
        "304":
          description: Not modified (matches the `If-None-Match` header)
...
  /internal/shadow-drift:
    get:
      operationId: getShadowDrift
      tags: [internal]
      summary: Shadow scoring drift report (reviewer-only, Task #639)
      description: |
        Reviewer-only listing of production reports whose shadow score
        diverges from the live score by ≥1 tier or ≥10 score points
        within the lookback window. Backs the "Shadow drift" tab on
        `/feedback-analytics`. Strict-auth: shadow drift is a leading
        signal of an in-flight scoring regression, same access policy
        as the other reviewer-only drift surfaces.
      parameters:
        - name: lookbackDays
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 14
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Shadow drift report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShadowDriftReport"
        "401":
          description: Reviewer token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ShadowDriftRow:
      type: object
      description: |
        One persisted shadow-score row whose live vs. shadow result
        diverges enough to surface on the reviewer drift dashboard
        (Task #639). A row qualifies when the tier flipped OR the
        score delta is at least 10 points in either direction.
      properties:
        id:
          type: integer
        reportId:
          type: integer
        liveScore:
          type: integer
        liveTier:
          type: string
        shadowScore:
          type: integer
        shadowTier:
          type: string
        scoreDiff:
          type: integer
          description: shadow_score − live_score (positive means shadow is harsher)
        tierDiverged:
          type: boolean
        shadowVersion:
          type: string
        scoredAt:
          type: string
          format: date-time
        fileName:
          type: string
          nullable: true
      required:
        - id
        - reportId
        - liveScore
        - liveTier
        - shadowScore
        - shadowTier
        - scoreDiff
        - tierDiverged
        - shadowVersion
        - scoredAt
    ShadowDriftTotals:
      type: object
      properties:
        total:
          type: integer
        divergent:
          type: integer
        tierFlips:
          type: integer
        scoreFlips:
          type: integer
        legitToSlop:
          type: integer
        slopToLegit:
          type: integer
      required:
        [total, divergent, tierFlips, scoreFlips, legitToSlop, slopToLegit]
    ShadowDriftReport:
      type: object
      description: |
        Reviewer-only listing of shadow-score divergences in the
        lookback window. `enabled` echoes whether
        `SHADOW_SCORING_ENABLED=1` is set on the API server so the UI
        can render an "OFF" hint when no fresh rows are landing.
      properties:
        enabled:
          type: boolean
        generatedAt:
          type: string
          format: date-time
        lookbackDays:
          type: integer
        scoreDeltaThreshold:
          type: integer
          description: Score-delta threshold (in points) above which a row is considered divergent regardless of tier.
        totals:
          $ref: "#/components/schemas/ShadowDriftTotals"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ShadowDriftRow"
      required:
        [enabled, generatedAt, lookbackDays, scoreDeltaThreshold, totals, rows]
...
    Webhook:
      type: object
      description: |
        A reviewer-registered webhook subscription. The plaintext signing
        secret is never returned by this schema — it is shown exactly
        once on the WebhookCreateResponse at registration time.
      properties:
        id:
          type: integer
        url:
          type: string
        eventTypes:
          type: array
          items:
            type: string
            enum: [report.scored]
        createdAt:
          type: string
          format: date-time
        lastDeliveredAt:
          type: string
          format: date-time
          nullable: true
        failureCount:
          type: integer
      required: [id, url, eventTypes, createdAt, lastDeliveredAt, failureCount]
    WebhookList:
      type: object
      properties:
        webhooks:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
      required: [webhooks]
    WebhookCreateBody:
      type: object
      properties:
        url:
          type: string
          description: Destination URL (http or https) up to 1000 characters.
        eventTypes:
          type: array
          description: |
            Subset of supported events to subscribe to. Defaults to
            `[report.scored]` when omitted. v1 only supports
            `report.scored`.
          items:
            type: string
            enum: [report.scored]
      required: [url]
    WebhookCreateResponse:
      allOf:
        - $ref: "#/components/schemas/Webhook"
        - type: object
          properties:
            secret:
              type: string
              description: |
                Per-webhook HMAC SHA-256 signing secret. Returned exactly
                once at registration time — only the SHA-256 hash is
                persisted server-side, so the caller must capture this
                value to verify incoming `X-VulnRap-Signature` headers.
                Re-register the webhook to receive a new secret.
          required: [secret]
    WebhookDeleteResponse:
      type: object
      properties:
        id:
          type: integer
      required: [id]
...